Add a lock-free list/hash table implementation
Use it in the mmap() MAP_ANONYMOUS implementation, which will allow it
to be AS-Safe and AC-Safe, as mmap() must be.

We also use a lock-free hash table to store thread pointers. Storing
thread pointers in a data structure as opposed to some special
register-like construct as is done on all other platforms, is likely
going to make TLS in general horrifically slow. However, at the
moment, we cannot think of a better solution. The ABI does not give us
any registers that we can use without any worry that they will be
clobbered if we call into foreign code. TLS must be AS-Safe and
AC-safe, so the data structure we used had to be lock-free.

Also add the start of a not-yet-functional z/OS futex implementation.
While a userspace emulation of futexes is not ideal, the alternative
is reimplementing everything in glibc using futexes in terms of
z/OS-native constructs, which which would be very tricky to get exactly
right and likely a major source of bugs for the forseeable future.