Move TLS implementation into rtld
Since TLS may be used within rtld, and rtld must be independent of libc,
we must move the entire TLS implementation into rtld. This includes the
lock-free list implementation. We make the thread pointer functions weak
inside of libc because it does not link against rtld at static link
time, so otherwise the link would fail.

Since the lock-free list implementation is also used elsewhere in libc,
right now we also export many of the symbols involved in that
implementation, which is not ideal.

INCU-180