.init_array/.fini_array fixes
Add in a workaround to .init_array and .fini_array processing that
prevents us from trying to execute null and -1 as functions. It's
unclear where the actual problem is, and if those sections should even
contain those values. We add in some code to check for those values and
ignore them.

INCU-175