Allow inline usages of rt_sigprocmask
rt_sigprocmask is used via the syscall macros in parts of the library
that we don't want to modify unless absolutely necessary. It's easy
enough for us to implement in terms of the regular sigprocmask, since
we don't have any rt signal support. In linux's aio_misc.h, it is used
inside an extern inline function, which breaks our entire syscall
system, so we need to override it with our own version.

INCU-227