Fixed __longjmp's return convention to match z/OS
Our implementation of __longjmp used the old Linux return register
convention, placing the return value in r2. z/OS expects the return
value to be placed in r15. This caused the setjmp call site always
return 1 when jumped to by longjmp. With this change the return now
behaves correctly, setting the return value to the argument unless it
is 0 in which case the return value becomes 1.

Since bash relies rather alot on setjmp/longjmp to handle evalution
errors this fixes many of those. Specficly those in which evaluation
is expected to continue after an error handled by a long jump.