Fix open's fd tagging behavior for pure ASCII file
When opening a pure text ASCII file it is important that the file
ccsid is marked ASCII. When the running program is 819 the file ccsid
has no effect.

However consider a program (like bash) which opens the ASCII file,
forks, dup that file and exec a 1047 program.  When the 1047 program
goes to read that file it will treat it according to the file
ccsid. Currently we have it marked as FT_BINARY, when in fact the file
is 819. This commit remedies this issue.