Changed pipe and open to surpress pipe tagging race
Fixed an issue in which open used the wrong stat stucture to determine
if a file was a pipe.

Changed pipe creation to always be 819 with no defer tag set.

Changing the pipe creation to not use FD_DEFER surpressed (and may or
may not have fixed) a race condition in which encoding of the below
command is sometimes 1047 and sometimes 819.
./bash -c "cat < <(echo hello)"

In particular if the echo won the race the output would be badly
encoded, so the below command would consistantly fail: ./bash -c
"(sleep 1; cat) < <(echo hello)"

And the following would consistantly succeed:
./bash -c "cat < <(sleep 1; echo hello)"

Removing FT_DEFER and explictly setting the tag to 819 on the pipe
creation didn't create any obvious issues in the bash test suite.

This applies to INCU-330.