Remove tagging of pipes on std{out, err} before main
If STDOUT or STDERR were a FIFO we would attempt to tag the underling
untagged pipe. This would create an issue when running bash (or any
toolchain compiled program) via ssh:

ssh user@mainframe /path/to/bash

The output of the above command would be ASCII which was assumed to be
EBCDIC and then converted from EBCDIC to ASCII. The checkfds code
would tag the untagged fifo and attempt to output ASCII. Then
somewhere between bash and the output terminal the EBCDIC to ASCII
conversion would occur. It is unclear, but doubtfull that the tagging
of the FIFO was semanticly correct in this case; clearly we violated
someone's expections around the output of that pipe. If instead we
don't attempt to tag the FIFO and respect the untagged value, and
output EBCDIC we don't have the ssh issue.

The bash test suit was run after this change and the tests all passed
or fail in and expected way.