Handle signal termination in the SIR
If we let the kernel handle signal default action termination, it prints
messages to the joblog, which is often stderr. These messages come out
in the wrong encoding much of the time, because the kernel doesn't
convert them to the encoding of the target stream. We can avoid these
unsightly messages by handling signal termination manually in the SIR.

USSP-1523