Expose an interface for using chattr/fchattr
We need to publicly expose some interface to use the chattr family of
z/OS-specific syscalls because they are the only interface for certain
operations. We have chosen to call the functions zos_chattr,
zos_fchattr, and zos_lchattr because unprefixed the functions are named
too simply and may reasonably be expected to conflict with with
identifiers used in user programs. For the same reason, we require that
_GNU_SOURCE be defined for the declarations to be exposed. The
declarations shall go in sys/stat.h, because they rely on the
declaration of zos_file_tag, which we have relocated into that header as
well so that it may be included in the definition of struct stat.

INCU-356