Implement 64-bit allocation, use it for brk
Add a C interface to the IARV64 service that allows allocation above the
2 GB bar. Furthermore, it supports a feature called 'guard areas' that
allows parts of the address space to be reserved without immediately
reserving storage for all of it. We use guard areas to implement a brk
function that acts similarly to how one might act on a regular unix
system, without an unacceptable memory overhead.

We will likely use IARV64 to implement anonymous mappings, but it is
harder to give it lazy allocation semantics in that situation.

INCU-355