Improve DIR interface such that fdopendir works
z/OS has two distinct notions of file descriptors where Linux make no
distinction. On z/OS there are File_Descriptors and
Directory_File_Descriptors, here onward refered to as fds and dfds
respectivly. A fd may be opened on a directory, however this does
not make it a dfd. Dfds must be opened with the z/OS native opendir.
Another restriction of dfds is they may not be dup()-ed or fcntl()-ed.
However fds opened on directories may be. Thus we do the following:
Never directly expose a dfd to any gcc program. Instead allow the use
of open() to create fds which refer to directories. These may be
fcntl()-ed and dup()-ed when the program wishes to use their fd
i.e. by way of fdopendir or opendir we use ioctl to retreave the path
associated with the fd and open a dfd on their behalf which may be used
to read the directory.

The ioctl for retreaving the path is also used to implement the *at
functions.

In order to get seekdir and telldir to work properly readdir will
need to be modified to used the readdir2 interface and the index
protocol.

Common functionality factored out.

INCU-373
INCU-364
INCU-372
INCU-386
INCU-401