Implement fchmodat, renameat and utimensat
With the addition of the get_path_at_fd_{cancel, nocancel} it is possible to
implement *at functions. These three are used by GNU patch and are therefore
useful to have implemented to continue development of other packages.

Note that some functionality around symlinks is not supported and there
isn't a clear way to support it.

Note that GNU patch uses the file type numbers directly. Thus the
z/OS specific numbers for __S_IF* should be hidden and the default
values should be used.

USSP-1150
INCU-387